+++ index.ts
@@ -1,1 +1,2 @@
+export * as ResponseEnvelopes from "./ResponseEnvelopes";
 export * as ProblemDetails from "./ProblemDetails";